pyproject.toml
[project]
name = "scrapy-playwright-scrapegraphai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "aiosql>=13.0",
    "logfire>=2.1.2",
    "polars>=1.12.0",
    "psycopg2-binary>=2.9.10",
    "pydantic>=2.9.2",
    "python-dotenv>=1.0.1",
    "pyyaml>=6.0.2",
    "scrapy-playwright>=0.0.42",
    "scrapy>=2.11.2",
]

[dependency-groups]
dev = [
    "ipykernel>=6.29.5",
]
